+++ R/bootstrap_parameters.R
@@ -19,10 +19,15 @@
 #'   bootstrapped coefficients. The resulting replicated for each coefficient
 #'   are treated as "distribution", and is passed to \code{\link[bayestestR:describe_posterior]{describe_posterior()}}
 #'   to calculate the related indices defined in the \code{"test"} argument.
+#'   \cr\cr
+#'   Note that that p-values returned here are "reversed" p-values - they are
+#'   the probability under the bootstrapped distribution of obtaining the null
+#'   (0) or more extreme.
 #'
 #' @examples
 #' \donttest{
 #' if (require("boot")) {
+#'   set.seed(2)
 #'   model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
 #'   b <- bootstrap_parameters(model)
 #'   print(b)